refactor: rename relevant code to label_studio (#233)
