+++ convert/src/label_studio_to_badegerdoc/label_studio_to_badgerdoc_use_case.py
@@ -8,10 +8,10 @@ from .badgerdoc_format.pdf_renderer import PDFRenderer
 from .badgerdoc_format.plain_text_converter import (
     TextToBadgerdocTokensConverter,
 )
-from .models.vertex_models import LabelStudioModel, S3Path
+from .models.label_studio_models import LabelStudioModel, S3Path
 
 
-class VertexToBDConvertUseCase:
+class LabelStudioToBDConvertUseCase:
     def __init__(
         self,
         s3_client,
