+++ convert/src/label_studio_to_badegerdoc/models/__init__.py
@@ -1,12 +1,12 @@
 from .bd_annotation_model import BadgerdocAnnotation, Obj, Size
 from .bd_tokens_model import BadgerdocToken, Offset, Page, PageSize
 from .common import S3Path
-from .vertex_models import (
+from .label_studio_models import (
     Annotation,
-    BadgerdocToVertexRequest,
+    BadgerdocToLabelStudioRequest,
     LabelStudioModel,
     ModelItem,
     Prediction,
     ResultItem,
-    VertexRequest,
+    LabelStudioRequest,
 )
