+++ convert/src/label_studio_to_badegerdoc/models/bd_annotation_model.py
