+++ convert/src/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI  # type: ignore
 
 from src.config import API_NAME, API_VERSION, settings
 from src.logger import get_logger
-from src.routers import coco, text, vertex
+from src.routers import coco, text, label_studio
 
 LOGGER = get_logger(__file__)
 
@@ -14,5 +14,5 @@ app = FastAPI(
     servers=[{"url": settings.root_path}],
 )
 app.include_router(coco.router)
-app.include_router(vertex.router)
+app.include_router(label_studio.router)
 app.include_router(text.router)
