+++ convert/tests/test_label_studio/test_data/badgerdoc_annotation.json
