+++ convert/tests/test_label_studio/test_data/input_text_field.json
