+++ convert/tests/test_label_studio/test_data/label_studio_format.json
